feat(auto-test): mobile table
